feat: js codehints initial  draft with webworkers
