--- src/JSUtils/worker/package.json
@@ -1,7 +1,0 @@
-{
-  "name": "brackets-javascript-code-hints",
-  "dependencies": {
-    "acorn": "5.1.1",
-    "tern": "0.21.0"
-  }
-}
